organizei o main
