+++ grafos/aula1.c
@@ -12,13 +12,34 @@ void preencher_matriz(int n, int matriz[n][n]){
 }
 
 void imprimirMatriz(int n, int matriz[n][n]) {
+    // 1. Imprime os índices das colunas (cabeçalho)
+    printf("     "); // Imprime um espaço inicial para alinhar com os índices das linhas
+    for (int j = 0; j < n; j++) {
+        printf("[%d]\t", j); // Imprime cada índice de coluna
+    }
+    printf("\n"); // Pula para a próxima linha
+
+    // Imprime uma linha de separação para melhor visualização
+    printf("----");
+    for (int j = 0; j < n; j++) {
+        printf("-----\t");
+    }
+    printf("\n");
+
+
+    // 2. Itera sobre as linhas para imprimir o índice da linha e os dados
     for (int i = 0; i < n; i++) {
+        // Imprime o índice da linha atual
+        printf("[%d] |\t", i); 
+        
+        // Imprime os elementos da matriz para a linha atual
         for (int j = 0; j < n; j++) {
             printf("%d\t", matriz[i][j]); // \t para alinhar as colunas
         }
-        printf("\n"); // Pula para a próxima linha
+        printf("\n"); // Pula para a próxima linha ao final de cada linha da matriz
     }
 }
+
 void copiarMatriz(int n, int matrizOriginal[n][n], int matrizCopia[n][n]){
     for(int l = 0; l < n; l++)
         for(int c = 0; c < n; c++){
@@ -58,63 +79,5 @@ void multiplicarMatrizes(int n, int matrizA[n][n], int matrizB[n][n], int matriz
         }
     }
 }
-int main() {
-    int n,u,v; // Variável para a dimensão da matriz
-
-    printf("--- Funções para Matrizes: Inicializar e Copiar ---\n\n");
-
-    // 1. Obter o tamanho da matriz
-    printf("Digite o tamanho N para a matriz quadrada (N x N): ");
-    scanf("%d", &n);
-
-    // Validação simples para evitar tamanhos negativos ou zero
-    if (n <= 0) {
-        printf("O tamanho da matriz deve ser um número positivo.\n");
-        return 1; // Termina o programa com erro
-    }
-
-    // 2. Declarar as matrizes
-    int matrizOriginal[n][n];
-    int matrizCopia[n][n];
-    int matrizResultado[n][n];
-
-    // 3. Chamar a função para preencher a matriz original com zeros
-    printf("\nInicializando a matriz original com zeros...\n");
-    inicializarComZeros(n, matrizOriginal);
-
-    // Imprimir a matriz original para verificar
-    printf("--- Matriz Original (após inicialização) ---\n");
-    imprimirMatriz(n, matrizOriginal);
-
-    printf("--Digite dois vertices que tem uma areasta entre eles--");
-    do{
-    scanf("%d %d", &u,&v);
-    colocarUm(n, matrizOriginal, u,v);
-    }while ((u != 0 || v != 0));    
-    
-    // ---------------------------------------------------------------------
-
-    // 4. Chamar a função para copiar a matriz original para a matrizCopia
-    printf("\nCopiando a matriz original para uma nova matriz...\n");
-    copiarMatriz(n, matrizOriginal, matrizCopia);
-
-    printf("--- Matriz Original ---\n");
-    imprimirMatriz(n, matrizOriginal);
-
-
-    // 5. Imprimir a matriz cópia para verificar
-    printf("--- Matriz Cópia ---\n");
-    imprimirMatriz(n, matrizCopia);
-
-
-    printf("--- Matriz multiplicada ---\n");
-    multiplicarMatrizes(n,matrizOriginal,matrizCopia,matrizResultado);
-    imprimirMatriz(n, matrizResultado);
-    
-  
-
-    return 0; // Sucesso
-}
-
